requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
boto3==1.15.15            # via -r requirements/base.in
boto==2.42.0              # via -r requirements/base.in
botocore==1.18.15         # via boto3, s3transfer
certifi==2020.6.20        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cryptography==3.1.1       # via django-fernet-fields, pyjwt
django-cors-headers==3.5.0  # via -r requirements/base.in
django-countries==6.1.3   # via -r requirements/base.in
django-crum==0.7.7        # via edx-rbac
django-fernet-fields==0.6  # via edx-enterprise-data
django-model-utils==4.0.0  # via edx-enterprise-data, edx-rbac
django-rest-swagger==2.2.0  # via -r requirements/base.in
django-storages==1.8      # via -c requirements/constraints.txt, -r requirements/base.in
django-waffle==2.0.0      # via edx-django-utils, edx-drf-extensions
django==2.2.16            # via -c requirements/constraints.txt, -r requirements/base.in, django-cors-headers, django-crum, django-fernet-fields, django-model-utils, django-storages, djangorestframework, drf-jwt, edx-django-release-util, edx-django-utils, edx-drf-extensions, edx-enterprise-data, edx-rbac, rest-condition
djangorestframework-csv==2.1.0  # via -r requirements/base.in
djangorestframework==3.12.1  # via -r requirements/base.in, django-rest-swagger, djangorestframework-csv, drf-jwt, edx-drf-extensions, rest-condition
drf-jwt==1.17.2           # via edx-drf-extensions
edx-ccx-keys==1.1.0       # via -r requirements/base.in
edx-django-release-util==0.4.4  # via -r requirements/base.in
edx-django-utils==3.8.0   # via -r requirements/base.in, edx-drf-extensions, edx-enterprise-data, edx-rest-api-client
edx-drf-extensions==6.2.0  # via -r requirements/base.in, edx-enterprise-data, edx-rbac
edx-enterprise-data==2.1.3  # via -r requirements/base.in
edx-opaque-keys==2.1.1    # via -r requirements/base.in, edx-ccx-keys, edx-drf-extensions, edx-enterprise-data
edx-rbac==1.3.3           # via edx-enterprise-data
edx-rest-api-client==5.2.1  # via -r requirements/base.in, edx-enterprise-data
elasticsearch-dsl==0.0.11  # via -c requirements/constraints.txt, -r requirements/base.in
elasticsearch==1.9.0      # via elasticsearch-dsl
future==0.18.2            # via pyjwkest
idna==2.10                # via requests
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
jmespath==0.10.0          # via boto3, botocore
markdown==2.6.6           # via -r requirements/base.in
markupsafe==1.1.1         # via jinja2
newrelic==5.20.1.150      # via edx-django-utils
openapi-codec==1.3.2      # via django-rest-swagger
ordered-set==3.1.1        # via -c requirements/constraints.txt, -r requirements/base.in
pbr==5.5.0                # via stevedore
psutil==5.7.2             # via edx-django-utils
pycparser==2.20           # via cffi
pycryptodomex==3.9.8      # via pyjwkest
pyjwkest==1.4.2           # via edx-drf-extensions
pyjwt[crypto]==1.7.1      # via drf-jwt, edx-rest-api-client
pymongo==3.11.0           # via edx-opaque-keys
python-dateutil==2.8.1    # via botocore, edx-drf-extensions, elasticsearch-dsl
python-memcached==1.59    # via -r requirements/base.in
pytz==2020.1              # via django
pyyaml==5.3.1             # via edx-django-release-util
requests==2.24.0          # via coreapi, edx-drf-extensions, edx-enterprise-data, edx-rest-api-client, pyjwkest, slumber
rest-condition==1.0.3     # via edx-drf-extensions
rules==2.2                # via edx-enterprise-data
s3transfer==0.3.3         # via boto3
semantic-version==2.8.5   # via edx-drf-extensions
simplejson==3.17.2        # via django-rest-swagger
six==1.15.0               # via cryptography, djangorestframework-csv, edx-ccx-keys, edx-django-release-util, edx-drf-extensions, edx-enterprise-data, edx-opaque-keys, edx-rbac, elasticsearch-dsl, pyjwkest, python-dateutil, python-memcached
slumber==0.7.1            # via edx-rest-api-client
sqlparse==0.4.1           # via django
stevedore==3.2.2          # via edx-django-utils, edx-opaque-keys
tqdm==4.50.2              # via -r requirements/base.in
unicodecsv==0.14.1        # via djangorestframework-csv
uritemplate==3.0.1        # via coreapi
urllib3==1.25.10          # via -r requirements/base.in, botocore, elasticsearch, requests
